requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.9.1
cryptography==1.8.1
Django==1.10
djangorestframework==3.6.2
idna==2.5
mysqlclient==1.3.10
packaging==16.8
pycparser==2.17
pyOpenSSL==16.2.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
